+++ docs/conf.py
@@ -27,8 +27,8 @@ master_doc = 'index'
 # Metadata
 project = u'agate'
 copyright = u'2016, Christopher Groskopf'
-version = '1.3.2'
-release = '1.3.2'
+version = '1.4.0'
+release = '1.4.0'
 
 exclude_patterns = ['_build']
 pygments_style = 'sphinx'